server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Pillow==10.0.1
numpy==1.26.0
scikit-learn==1.3.1
opencv-python==4.8.1.78
torch==2.1.0  # Adjust version if needed based on your CUDA setup
sqlalchemy==2.0.21
psycopg2-binary==2.9.8
plotly==5.17.0

